Add new classes for address service batch processing
Added new classes to support processing of the address in batch. These classes include the model classes for cities and hromadas, handling CSV input files and saving entities, configuration classes that include bean definitions, repositories for querying the databases, and mappers for converting DTOs into Entities and vice versa. This will allow cities and hromadas data to be processed in bulk, improving efficiency and performance.